Lectures/chat-gpt-talks/gpt_code/requirements.txt
rich
numpy
opencv-python
pillow